+++ Back-End/src/main/java/com/ecoprint/control_center/repository/TipoImpressoraRepository.java
@@ -4,7 +4,7 @@ import com.ecoprint.control_center.model.TipoImpressora;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface TipoImpressoraRepository extends JpaRepository<TipoImpressora, Long> {
+public interface TipoImpressoraRepository extends JpaRepository<TipoImpressora, Integer> {
